+++ utils/getArrayData/index.ts
@@ -1,0 +1,1 @@
+export { getArrayData } from "./getArrayData";
