+++ utils/getStringData/index.ts
@@ -1,0 +1,1 @@
+export { getStringData } from "./getStringData";
